feat: add program data module and integrate it into Program component
