+++ src/components/Programs.js
@@ -1,6 +1,8 @@
 // src/components/Program.js
 import React, { useState, useEffect } from 'react';
 import './Program.css';
+import programData from '../data/programData'; // Import the program data
+
 
 const Program = () => {
   const [programs, setPrograms] = useState([]);
@@ -29,7 +31,9 @@ const Program = () => {
     fees: '',
     description: '',
   });
-
+  useEffect(() => {
+    setPrograms(programData); // Set the program data
+  }, []);
   // Check for logged-in user credentials and status
   useEffect(() => {
     const userCredentials = JSON.parse(localStorage.getItem('userCredentials')) || [];
@@ -41,43 +45,6 @@ const Program = () => {
         setIsAdmin(user.status === 'admin');
       }
     }
-  }, []);
-
-  useEffect(() => {
-    // Sample program data
-    const programData = [
-      {
-        id: 1,
-        code: 'SD-DIPLOMA',
-        name: 'Software Development - Diploma',
-        term: 'Winter',
-        startDate: 'September 5, 2024',
-        endDate: 'June 15, 2026',
-        fees: 'Domestic: $9,254 / International: $27,735',
-        description: 'A comprehensive two-year software development diploma program...',
-      },
-      {
-        id: 2,
-        code: 'SD-POST-DIPLOMA',
-        name: 'Software Development - Post-Diploma',
-        term: 'Winter',
-        startDate: 'September 5, 2024',
-        endDate: 'June 15, 2025',
-        fees: 'Domestic: $7,895 / International: $23,675',
-        description: 'Jumpstart your tech career with our one-year post-diploma program...',
-      },
-      {
-        id: 3,
-        code: 'SD-CERTIFICATE',
-        name: 'Software Development - Certificate',
-        term: 'Fall',
-        startDate: 'September 1, 2024',
-        endDate: 'February 28, 2025',
-        fees: 'Domestic: $5,000 / International: $15,000',
-        description: 'A short-term program to develop foundational skills in software...',
-      },
-    ];
-    setPrograms(programData);
   }, []);
 
   const handleProgramSelection = (program) => {
