+++ src/data/programData.js
@@ -1,0 +1,35 @@
+// src/data/programData.js
+const programData = [
+  {
+    id: 1,
+    code: 'SD-DIPLOMA',
+    name: 'Software Development - Diploma',
+    term: 'Winter',
+    startDate: 'September 5, 2024',
+    endDate: 'June 15, 2026',
+    fees: 'Domestic: $9,254 / International: $27,735',
+    description: 'A comprehensive two-year software development diploma program...',
+  },
+  {
+    id: 2,
+    code: 'SD-POST-DIPLOMA',
+    name: 'Software Development - Post-Diploma',
+    term: 'Winter',
+    startDate: 'September 5, 2024',
+    endDate: 'June 15, 2025',
+    fees: 'Domestic: $7,895 / International: $23,675',
+    description: 'Jumpstart your tech career with our one-year post-diploma program...',
+  },
+  {
+    id: 3,
+    code: 'SD-CERTIFICATE',
+    name: 'Software Development - Certificate',
+    term: 'Fall',
+    startDate: 'September 1, 2024',
+    endDate: 'February 28, 2025',
+    fees: 'Domestic: $5,000 / International: $15,000',
+    description: 'A short-term program to develop foundational skills in software...',
+  },
+];
+
+export default programData;
